+++ config/filament-starter.php
@@ -48,4 +48,33 @@ return [
     'superadmin' => [
         'role' => 'super_admin',
     ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Developer Gate
+    |--------------------------------------------------------------------------
+    */
+    'developer_gate' => [
+        'enabled' => env('STARTER_DEVELOPER_GATE_ENABLED', false),
+        'password' => env('STARTER_DEVELOPER_GATE_PASSWORD', 'secret'),
+    ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | Plugin Global Configurations
+    |--------------------------------------------------------------------------
+    */
+    'plugins' => [
+        'backgrounds' => [
+            'show_attribution' => true,
+            'remember' => 900,
+            'image_provider' => 'curated', // curated, my-images, triangles
+            'my_images_directory' => 'images/backgrounds',
+        ],
+        'revive' => [
+            'user_scoping' => true,
+            'tenant_scoping' => true,
+            'global_admin_panels' => ['admin'],
+        ],
+    ],
 ];